Remove unneeded copying.
When algorithm operates by modifying the input array in place, it is
good manners to make a copy first so that the function does not mutate
user input. I am guessing that that was the motivation for the use of
`copy()` here.

However, as far as I can tell, this algorithm does not modify `signal`
(or its alias `a`) in place. It reassigns the variable `a` in a loop,
but does not actually modify the input array object.

Therefore, we can remove this copying and increase the speed. Also, note
that the call to `np.array` explicitly cast the input to a literal numpy
array. By removing that from the code, we accept dask arrays and cupy
arrays and allow them to flow through the algorithm via the NEP-18 numpy
dispatch mechanism.